# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Import Contacts Endpoints",
        "description": "Import Contacts",
        "termsOfService": "http://github.com/terms/",
        "contact": {
            "email": "kiramishima@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/file/{filename}": {
            "get": {
                "tags": [
                    "Upload",
                    "CSV"
                ],
                "summary": "Allow upload a CSV File",
                "description": "Allow upload a CSV File that contain all the information.",
                "operationId": "GetFileInfo",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "authentication": []
                    }
                ]
            }
        },
        "/api/v1/upload_csv": {
            "post": {
                "tags": [
                    "Upload",
                    "CSV"
                ],
                "summary": "Allow upload a CSV File",
                "description": "Allow upload a CSV File that contain all the information.",
                "operationId": "UploadCSV",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "csvfile": {
                                        "description": "CSV File.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/file/{filename}/save": {
            "post": {
                "tags": [
                    "Upload",
                    "CSV"
                ],
                "summary": "Allow upload a CSV File",
                "description": "Allow upload a CSV File that contain all the information.",
                "operationId": "SaveReMap",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "remap": {
                                        "description": "Info for Remap",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "status": {
                        "description": "Status",
                        "type": "boolean"
                    },
                    "message": {
                        "description": "Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Success": {
                "properties": {
                    "status": {
                        "description": "Status",
                        "type": "boolean"
                    },
                    "data": {
                        "description": "Data",
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    }
}